Finished improving unit tests
